packetHandler class

+++ src/tools.h
@@ -7,8 +7,8 @@ typedef unsigned short int uint16;
 typedef short int sint16;
 typedef unsigned int uint32;
 typedef int sint32;
-typedef unsigned long int uint64;
-typedef long int sint64;
+typedef unsigned long long int uint64;
+typedef long long int sint64;
 
 void putSint64(uint8 *buf, long long value);
 void putUint32(uint8 *buf, uint32 value);
